+++ src/constants/api.js
@@ -1,3 +1,5 @@
 export const API_BASE = import.meta.env.VITE_API_BASE ?? 'http://localhost:3001'
 export const ADMIN_API_BASE = import.meta.env.VITE_ADMIN_API_BASE ?? 'http://localhost:8080'
 export const HR_API_BASE = import.meta.env.VITE_HR_API_BASE ?? 'http://localhost:8081'
+export const SCM_API_BASE = import.meta.env.VITE_SCM_API_BASE ?? 'http://localhost:8082'
+export const KMS_API_BASE = import.meta.env.VITE_KMS_API_BASE ?? 'http://localhost:8083'
