+++ src/services/hrmanager/promotionApi.js
@@ -1,4 +1,4 @@
-import hrApi from '../hrApi.js'
+import { hrApi } from '../apiClient'
 
 const promotionApi = {
   getSummary() {
