+++ src/services/knowledgeArticleApi.js
@@ -1,88 +1,6 @@
-import axios from 'axios'
-import { ADMIN_API_BASE } from '@/constants'
-import { useAuthStore } from '@/stores/auth'
-
-const KMS_API_BASE = import.meta.env.VITE_KMS_API_BASE ?? 'http://localhost:8083'
-
-const kmsApi = axios.create({
-  baseURL: KMS_API_BASE,
-  timeout: 5000,
-  withCredentials: true,
-})
-
-// ── request interceptor: Bearer 토큰 자동 첨부 ─────────────────
-kmsApi.interceptors.request.use((config) => {
-  const authStore = useAuthStore()
-  if (authStore.accessToken) {
-    config.headers.Authorization = `Bearer ${authStore.accessToken}`
-  }
-  return config
-})
-
-// ── response interceptor: 401 시 토큰 갱신 후 재시도 (authApi.js와 동일 패턴) ─
-let isRefreshing = false
-let failedQueue = []
-
-function processQueue(error, token) {
-  failedQueue.forEach(({ resolve, reject }) => {
-    if (error) reject(error)
-    else resolve(token)
-  })
-  failedQueue = []
-}
-
-kmsApi.interceptors.response.use(
-  (response) => response,
-  async (error) => {
-    const originalRequest = error.config
-
-    if (error.response?.status !== 401 || originalRequest._retry) {
-      return Promise.reject(error)
-    }
-
-    if (isRefreshing) {
-      return new Promise((resolve, reject) => {
-        failedQueue.push({ resolve, reject })
-      }).then((token) => {
-        originalRequest.headers.Authorization = `Bearer ${token}`
-        return kmsApi(originalRequest)
-      })
-    }
-
-    originalRequest._retry = true
-    isRefreshing = true
-
-    try {
-      const res = await axios.post(
-        `${ADMIN_API_BASE}/api/v1/auth/refresh`,
-        null,
-        { withCredentials: true }
-      )
-
-      if (res.data?.success && res.data.data?.accessToken) {
-        const newToken = res.data.data.accessToken
-        const authStore = useAuthStore()
-        authStore.setToken(newToken)
-        processQueue(null, newToken)
-        originalRequest.headers.Authorization = `Bearer ${newToken}`
-        return kmsApi(originalRequest)
-      }
-    } catch (refreshError) {
-      processQueue(refreshError, null)
-      const authStore = useAuthStore()
-      authStore.clearAuth()
-      window.location.href = '/login'
-      return Promise.reject(refreshError)
-    } finally {
-      isRefreshing = false
-    }
-
-    return Promise.reject(error)
-  }
-)
+import { kmsApi } from '@/services/apiClient'
 
 // ── API 메서드 ──────────────────────────────────────────────────
-
 const knowledgeArticleApi = {
 
   // ── 조회 ──────────────────────────────────────────────────────
